+++ scripts/common.js
@@ -1,0 +1,124 @@
+import {
+  sampleRUM,
+  loadCSS,
+  loadBlock,
+  loadBlocks,
+  loadFooter,
+  loadHeader,
+} from './lib-franklin.js';
+
+let placeholders = null;
+
+export async function getPlaceholders() {
+  placeholders = await fetch('/placeholder.json').then((resp) => resp.json());
+}
+
+export function getTextLabel(key) {
+  return placeholders.data.find((el) => el.Key === key)?.Text || key;
+}
+
+/**
+ * Create an element with the given id and classes.
+ * @param {string} tagName the tag
+ * @param {Object} options the element options
+ * @param {string[]|string} [options.classes=[]] the class or classes to add
+ * @param {Object} [options.props={}] any other attributes to add to the element
+ * @returns {HTMLElement} the element
+ */
+export function createElement(tagName, options = {}) {
+  const { classes = [], props = {} } = options;
+  const elem = document.createElement(tagName);
+  const isString = typeof classes === 'string';
+  if (classes || (isString && classes !== '') || (!isString && classes.length > 0)) {
+    const classesArr = isString ? [classes] : classes;
+    elem.classList.add(...classesArr);
+  }
+  if (!isString && classes.length === 0) elem.removeAttribute('class');
+
+  if (props) {
+    Object.keys(props).forEach((propName) => {
+      const value = propName === props[propName] ? '' : props[propName];
+      elem.setAttribute(propName, value);
+    });
+  }
+
+  return elem;
+}
+/**
+ * Adds the favicon.
+ * @param {string} href The favicon URL
+ */
+function addFavIcon(href) {
+  const link = createElement('link', { props: { rel: 'icon', type: 'image/svg+xml', href } });
+  const existingLink = document.querySelector('head link[rel="icon"]');
+  if (existingLink) {
+    existingLink.parentElement.replaceChild(link, existingLink);
+  } else {
+    document.getElementsByTagName('head')[0].appendChild(link);
+  }
+}
+
+export async function loadTemplate(doc, templateName) {
+  try {
+    const cssLoaded = new Promise((resolve) => {
+      loadCSS(`${window.hlx.codeBasePath}/templates/${templateName}/${templateName}.css`, resolve);
+    });
+    const decorationComplete = new Promise((resolve) => {
+      (async () => {
+        try {
+          const mod = await import(`../templates/${templateName}/${templateName}.js`);
+          if (mod.default) {
+            await mod.default(doc);
+          }
+        } catch (error) {
+          // eslint-disable-next-line no-console
+          console.log(`failed to load module for ${templateName}`, error);
+        }
+        resolve();
+      })();
+    });
+    await Promise.all([cssLoaded, decorationComplete]);
+  } catch (error) {
+    // eslint-disable-next-line no-console
+    console.log(`failed to load block ${templateName}`, error);
+  }
+}
+
+/**
+ * loads everything that doesn't need to be delayed.
+ */
+export async function loadLazy(doc) {
+  const main = doc.querySelector('main');
+  await loadBlocks(main);
+
+  const { hash } = window.location;
+  const element = hash ? doc.getElementById(hash.substring(1)) : false;
+  if (hash && element) element.scrollIntoView();
+  const header = doc.querySelector('header');
+
+  loadHeader(header);
+  loadFooter(doc.querySelector('footer'));
+
+  const subnav = header?.querySelector('.block.sub-nav');
+  if (subnav) {
+    loadBlock(subnav);
+  }
+
+  loadCSS(`${window.hlx.codeBasePath}/styles/lazy-styles.css`);
+  addFavIcon(`${window.hlx.codeBasePath}/styles/favicon.svg`);
+  sampleRUM('lazy');
+  sampleRUM.observe(main.querySelectorAll('div[data-block-name]'));
+  sampleRUM.observe(main.querySelectorAll('picture > img'));
+}
+
+/**
+ * loads everything that happens a lot later, without impacting
+ * the user experience.
+ */
+export function loadDelayed() {
+  // eslint-disable-next-line import/no-cycle
+  window.setTimeout(() => {
+    import('./delayed.js');
+  }, 3000);
+  // load anything that can be postponed to the latest here
+}
